Adds phpcs and resolves phpcs errors (#77)
* Adds phpcs and resolves phpcs errors

* oops

* move github actions into single ci file

---------
