+++ tests/TestCase/Meta/RequestMetadataTest.php
@@ -27,7 +27,7 @@ class RequestMetadataTest extends TestCase
         $request->expects($this->once())->method('clientIp')->will($this->returnValue('12345'));
         $request->expects($this->once())->method('getRequestTarget')->will($this->returnValue('/things?a=b'));
         $logs[] = new AuditDeleteEvent('1234', 1, 'articles');
-        $event = $this->dispatchEvent('AuditStash.beforeLog', ['logs' => $logs]);
+        $this->dispatchEvent('AuditStash.beforeLog', ['logs' => $logs]);
 
         $expected = [
             'ip' => '12345',
